Refactor + uncommented RequestBody tests
